+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 ### Bug fixes:
 
 - Pin `langchain-openai` to version `0.3.35` for compatibility with `typing_extensions<4.13` as requested by Plone `6.0.15` constraints. @arybakov05
+- Fixed tests for `helper` file. @arybakov05
 
 ## 2.0.0 (2026-02-13)